move verifyWorkspace to resolve circular dependency
