+++ libs/vscode/nx-workspace/src/index.ts
@@ -1,3 +1,4 @@
 export * from './lib/find-workspace-json-target';
 export * from './lib/reveal-workspace-json';
 export * from './lib/workspace-codelens-provider';
+export * from './lib/verify-workspace';
