+++ libs/vscode/nx-workspace/src/lib/verify-workspace.ts
