+++ libs/vscode/tasks/src/lib/cli-task-provider.ts
@@ -8,8 +8,8 @@ import {
 } from 'vscode';
 
 import { getTelemetry } from '@nx-console/server';
-// eslint-disable-next-line @nrwl/nx/enforce-module-boundaries
-import { verifyWorkspace, verifyNodeModules } from '@nx-console/vscode/verify';
+import { verifyWorkspace } from '@nx-console/vscode/nx-workspace';
+import { verifyNodeModules } from '@nx-console/vscode/verify';
 import { CliTask } from './cli-task';
 import {
   CliTaskDefinition,
