+++ libs/vscode/verify/src/index.ts
@@ -1,3 +1,2 @@
 export * from './lib/verify-builder-definition';
 export * from './lib/verify-node-modules';
-export * from './lib/verify-workspace';
